reach into freerdp and check ssl buffer for data
this is a horrible hack, but it fixes the honeypot reset race condition
in guac only, without resorting to a freerdp change.

obviously if the struct layout changes this will break :/

There is probably a bug fix that needs to happen deep inside libfreerdp,
but as this doesnt seem to happen for any other RDP client it is not
really worth the investment.